+++ goparser/__init__.py
@@ -1,3 +1,7 @@
+from __future__ import (absolute_import, division,
+                        print_function, unicode_literals)
+from builtins import *
+
 import pkg_resources
 
 from goparser.term import GOTerm