+++ goparser/annotation.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2015 Florian Wagner
+# Copyright (c) 2015, 2016 Florian Wagner
 #
 # This file is part of GOparser.
 #
@@ -14,10 +14,17 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import re
+"""Module containing the `GOAnnotation` class."""
+
+from __future__ import (absolute_import, division,
+                        print_function, unicode_literals)
+from builtins import *
+
+# import re
 from collections import Iterable
 
 from goparser import GOTerm
+
 
 class GOAnnotation(object):
 
@@ -74,80 +81,80 @@ class GOAnnotation(object):
 
     """
 
-    _evidence_name = {\
-            'EXP': 'experiment',\
-            'IDA': 'direct assay',\
-            'IPI': 'physical interaction',\
-            'IMP': 'mutant phenotype',\
-            'IGI': 'genetic interaction',\
-            'IEP': 'expression pattern',\
-            'ISS': 'sequence or structural similarity',\
-            'ISO': 'sequence orthology',\
-            'ISA': 'sequence alignment',\
-            'ISM': 'sequence model',\
-            'IGC': 'genomic context',\
-            'IBA': 'biological aspect of ancestor',\
-            'IBD': 'biological aspect of descendant',\
-            'IKR': 'key residues',\
-            'IRD': 'rapid divergence',\
-            'RCA': 'reviewed computational analysis',\
-            'TAS': 'traceable author statement',\
-            'NAS': 'non-traceable author statement',\
-            'IC' : 'inferred by curator',\
-            'ND' : 'no biological data available',\
-            'IEA': 'inferred from electronic annotation'\
-            }
+    _evidence_name = {
+        'EXP': 'experiment',
+        'IDA': 'direct assay',
+        'IPI': 'physical interaction',
+        'IMP': 'mutant phenotype',
+        'IGI': 'genetic interaction',
+        'IEP': 'expression pattern',
+        'ISS': 'sequence or structural similarity',
+        'ISO': 'sequence orthology',
+        'ISA': 'sequence alignment',
+        'ISM': 'sequence model',
+        'IGC': 'genomic context',
+        'IBA': 'biological aspect of ancestor',
+        'IBD': 'biological aspect of descendant',
+        'IKR': 'key residues',
+        'IRD': 'rapid divergence',
+        'RCA': 'reviewed computational analysis',
+        'TAS': 'traceable author statement',
+        'NAS': 'non-traceable author statement',
+        'IC' : 'inferred by curator',
+        'ND' : 'no biological data available',
+        'IEA': 'inferred from electronic annotation'
+    }
     """Mapping of the three-letter evidence codes to their full names.
     """
 
-    _evidence_type = {\
-            'EXP': 'experimental',\
-            'IDA': 'experimental',\
-            'IPI': 'experimental',\
-            'IMP': 'experimental',\
-            'IGI': 'experimental',\
-            'IEP': 'experimental',\
-            'ISS': 'computational',\
-            'ISO': 'computational',\
-            'ISA': 'computational',\
-            'ISM': 'computational',\
-            'IGC': 'computational',\
-            'IBA': 'computational',\
-            'IBD': 'computational',\
-            'IKR': 'computational',\
-            'IRD': 'computational',\
-            'RCA': 'computational',\
-            'TAS': 'literature',\
-            'NAS': 'literature',\
-            'IC' : 'curator',\
-            'ND' : 'no_data',\
-            'IEA': 'automatic'\
-            }
+    _evidence_type = {
+        'EXP': 'experimental',
+        'IDA': 'experimental',
+        'IPI': 'experimental',
+        'IMP': 'experimental',
+        'IGI': 'experimental',
+        'IEP': 'experimental',
+        'ISS': 'computational',
+        'ISO': 'computational',
+        'ISA': 'computational',
+        'ISM': 'computational',
+        'IGC': 'computational',
+        'IBA': 'computational',
+        'IBD': 'computational',
+        'IKR': 'computational',
+        'IRD': 'computational',
+        'RCA': 'computational',
+        'TAS': 'literature',
+        'NAS': 'literature',
+        'IC' : 'curator',
+        'ND' : 'no_data',
+        'IEA': 'automatic'
+    }
     """Mapping of the three-letter evidence codes to their evidence types.
     """
 
-    _evidence_type_short = {\
-            'experimental': 'exp.',\
-            'computational': 'comp.',\
-            'literature': 'lit.',\
-            'curator': 'cur.',\
-            'no_data': 'n.d.',\
-            'automatic': 'autom.'\
-            }
+    _evidence_type_short = {
+        'experimental': 'exp.',
+        'computational': 'comp.',
+        'literature': 'lit.',
+        'curator': 'cur.',
+        'no_data': 'n.d.',
+        'automatic': 'autom.'
+    }
     """Mapping of the evidence types to abbreviated forms.
     """
 
-    #uniprot_pattern = re.compile("([A-Z][A-Z0-9]{5})(?:-(\d+))?")
+    # uniprot_pattern = re.compile("([A-Z][A-Z0-9]{5})(?:-(\d+))?")
 
-    def __init__(self, gene, term, evidence, db_id = None,
-                db_ref = None, with_ = None):
+    def __init__(self, gene, term, evidence, db_id=None,
+                 db_ref=None, with_=None):
 
         assert isinstance(term, GOTerm)
-        assert isinstance(gene, (str, unicode)) and gene != ''
-        assert isinstance(evidence, (str, unicode)) and evidence != ''
+        assert isinstance(gene, str) and gene != ''
+        assert isinstance(evidence, str) and evidence != ''
 
         if db_id is not None:
-            assert isinstance(db_id, (str, unicode)) and db_id != ''
+            assert isinstance(db_id, str) and db_id != ''
 
         if db_ref is not None:
             assert isinstance(db_ref, Iterable)
@@ -165,13 +172,13 @@ class GOAnnotation(object):
         self.with_ = () if with_ is None else tuple(with_)
 
     def __repr__(self):
-        return '<GOAnnotation (hash=%d)>' %(hash(self))
+        return '<GOAnnotation (hash=%d)>' % hash(self)
 
     def __str__(self):
         return '<GOAnnotation of gene "%s" with term "%s" (%s)>' \
-                %(self.gene, self.term.name, self.term.id)
+                % (self.gene, self.term.name, self.term.id)
 
-    def __eq__(self,other):
+    def __eq__(self, other):
         if type(self) != type(other):
             return False
         elif self is other:
@@ -180,21 +187,20 @@ class GOAnnotation(object):
             return repr(self) == repr(other)
 
     def __hash__(self):
-        data = []
-        data.append(self.gene)
-        data.append(self.term)
-        data.append(self.evidence)
-        data.append(self.db_id)
-        data.append(self.db_ref)
-        
-        return hash(tuple(data))
+        data = (
+            self.gene,
+            self.term,
+            self.evidence,
+            self.db_id,
+            self.db_ref,
+        )
+        return hash(data)
 
     def get_gaf_format(self):
         """Return a GAF 2.0-compatible string representation of the annotation.
 
         Parameters
         ----------
-        None
 
         Returns
         -------
@@ -202,10 +208,11 @@ class GOAnnotation(object):
             The formatted string.
         """
         sep = '\t'
-        return sep.join([self.gene, self.db_ref, self.term.id, self.evidence,
-                '|'.join(self.db_ref), '|'.join(self.with_)])
+        return sep.join(
+            [self.gene, self.db_ref, self.term.id, self.evidence,
+             '|'.join(self.db_ref), '|'.join(self.with_)])
 
-    #def get_pretty_format(self):
+    # def get_pretty_format(self):
     #    """Returns a nicely formatted string with the annotation information.
     #
     #    Parameters
@@ -217,6 +224,8 @@ class GOAnnotation(object):
     #    str
     #        The formatted string.
     #    """
-    #    pretty = "Annotation of gene '%s' with GO term '%s' (%s, reference: %s)'" \
-    #            %(self.target,self.term.get_pretty_format(),self.evidence,'|'.join(self.db_ref))
+    #    pretty = 'Annotation of gene "%s" with GO term "%s"
+    #              (%s, reference: %s)'" \
+    #              %(self.target,self.term.get_pretty_format(),self.evidence,
+    #               '|'.join(self.db_ref))
     #    return pretty